refactor(rolldown_plugin): extract native plugin context
