+++ crates/rolldown_plugin/src/plugin_context/mod.rs
@@ -1,5 +1,7 @@
+mod native_context;
 mod plugin_context;
 mod transform_plugin_context;
 
-pub use plugin_context::{PluginContext, PluginContextImpl};
+pub use native_context::PluginContextImpl;
+pub use plugin_context::PluginContext;
 pub use transform_plugin_context::{SharedTransformPluginContext, TransformPluginContext};
